refactor: change location of components
